+++ config/vim/pack/plugins/start/fzf/CHANGELOG.md
@@ -8,6 +8,9 @@ CHANGELOG
   # Display "Type to search" when the input is empty
   fzf --ghost "Type to search"
   ```
+- Added `change-ghost` and `transform-ghost` actions for dynamically changing the ghost text
+- Added `change-pointer` and `transform-pointer` actions for dynamically changing the pointer sign
+- Added `r` flag for placeholder expression (raw mode) for unquoted output
 - Bug fixes and improvements
 
 0.60.3